CR-70508: Explicitly stop glib processing on server destroy
When shutting down a server being processed with glib event loops
we should explicitly stop the processing of accepted connections
otherwise there is no guarantee when/if the sockets will be closed.
This is because we only close the socket and remove the glib source
if the client closes its end and the glib event loop notices this
activity on the socket. Furthermore this could lead to a use after
free as by this stage the server structure has already been destroyed.

When closing all accepted connections we need to be careful not to
close the accept thread eventfd that is being processed.

Change-Id: Ica161db745f525141a8127c682dd224df0643530
Reviewed-by: Thomas Winter <thomas.winter@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>