CR-70827: Send remote host deletion event to publishers
Rather than sending multiple singular removal events for each method
subscription that was on a remote host send a single update message
to the publishers for the service. The publisher can then delete all
of the subscriptions that were from the now removed host. Furthermore,
there is no need for the cmsg_psd daemon to synchronously wait for this
operation to complete.

This improves the deletion time for the subscriptions, reduces the
number of individual messages that must be sent, and ensures that
the cmsg_psd daemon does not wait large amounts of time for this
operation to complete.

Change-Id: I021aa7db6f5e8807a7e88ced7323c42bdbbfce32
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>