CR-70771: Fix peeking for packet header on TCP transports
The 'recv_all' function used by the TCP transport does not correctly
peek the data if the full length of data is not already on the socket.
For example, if we are expecting 20 bytes, but only 10 bytes are present
on the first peek then we will read the first 10 bytes twice (as MSG_PEEK
does not remove the data from the socket), resulting in incorrect data
being returned from the call.

Switch the TCP transport to use the same recv function as the UNIX and
TIPC transports while adding support to this function for receiving in
chunks. This function correctly peeks (implicitly) via the MSG_DONTWAIT
path.

Change-Id: I0dcf3879e9811b11d66a1f61d427ce65ac53bffe
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>