ENGBL-1307: Cache TCP connection info to avoid unnecessary connect attempts
The default connection timeout of TCP (and TIPC) connections is 5 seconds.
When stack members abruptly leave the stack a number of CMSG clients/publishers
attempt to reconnect to the now gone member resulting in multiple connection
attempts that simply block for 5 seconds before failing. We can improve this
by caching the stack member addresses and their status in shared memory resulting
in instant failure when attempting to connect to a stack member that has left.

Note that the cache size is arbitrarily set to 24 which is the maximum number of
nodes in a fully loaded VCS+ setup (and we currently only care about optimising
the TCP usage for stacking). This value of 24 matches the maximum number of clients
allowed in a composite client.

Change-Id: Icb6fe26a849d4703e6ee87ef12e8bc3c68995174
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>