TORUSAPI-569: Remove duplication in auto-generated CMSG API code
The CMSG API code is changed to use static inlines for API functions with
a single call. This reduces the size of the API libraries significantly.

The API code that was previously autogenerated is moved to cmsg_client.c
and modified to all go through a single function. The optional extensions
(file response and service support check) are handled by adding data to
a new cmsg_api_descriptor that wraps the protobuf descriptor.

The libraries used by cmsg proxy are changed to include the cmsg descriptor
and method index so that cmsg_api_invoke can be called directly. This
simplifies the code to call the API in cmsg proxy. The input and output
descriptor pointers are removed as they are accessible via cmsg descriptor.
Another reason for this change is to avoid adding pointers to
static inlines to the proxy def library.

Unit tests are updated. The cmsg_api_invoke function has been modified to
be able to selectively mock it for only certain calls (can still call the
real version if mocking)

Change-Id: Ic1e2c938268870e6a8098f139b96d6f2b3a845fc
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>