TORUSAPI-569: Remove duplication in auto-generated CMSG IMPL code
The CMSG server code is changed to use a generic function to invoke the
impl functions. This is done by defining a new generic structure for a
cmsg service that takes the list of impl functions and any extensions as
an array of pointers.

Instead of calling protobuf_c_service_invoke_internal to invoke the API,
a new function cmsg_server_invoke_impl is used instead. This includes all
of the same checks as the old function, but calls the API in the new way,
including any pre-impl extensions like validation.

The impl pointers int the auto-generated code are cast to void to handle
functions with/without an input parameter in the same way. They are cast
back to the correct type in cmsg_server_call_impl before calling.

cmsg_method_extensions is renamed to cmsg_method_client_extensions to make
it more obviously different from the new server extensions.

Change-Id: Ie4082b18db9cb45ba901b366553ba2f9295a3758
Reviewed-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>