ENGBL-1307: Reset TCP connections when closing sockets
When shutting down and closing a TCP socket the kernel will attempt
to finish the connection as specified by the TCP protocol. However, if
the other end of the connection has left, and never comes back, then the
kernel will keep the sockets around internally for a long time waiting to
complete the protocol. To avoid resources being left around we choose to
forcibly reset the connection (via SO_LINGER) to ensure the resources are
released immediately when the socket is closed. We do this for all accepted
server sockets and the client socket for 2-way clients.

For 1-way clients we cannot hard reset the connection as we have no guarantee
that the kernel has fully drained the socket. Therefore we allow the kernel
to attempt to gracefully shutdown the connection (and thus fully flush the
socket). However, we enable keepalive probes on the connection so that it
will promptly detect the case where the other end has disappeared and reset
the connection.

Change-Id: I792178d5c8b86cc2f92af26da0da0b9577a811af
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>