ENGBL-1307: Improve thread cancellation logic exposed in tests
Sporadically the functional tests are failing due to a socket leak.
From debugging it was found there were two possible places this could
occur.

- The cancellation points around the cmsg_server_accept call were not
inclusive enough meaning that we could cancel just after accepting the
socket and leak it.
- The logic in '_clear_accept_sd_queue' is wrong. The queue stores pointers
to integers, not integers stored in a pointer value.

This code has not been modified in a while, however recent changes in
the library seem to have caused the tests to fail more frequently.

Change-Id: I24e1de839919ea143eeaac7e34ddc0d6aaf132f0
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>