CR-71767: Fix deadlock on accept thread cancellation
There are a couple of issues with the current select() usage in this function:

1. We are not setting the read_fds variable each time before calling select() (it is
modified after each call).

2. We don't check the return value of select.

This means that currently the select call can return when there is actually
no connection to accept. Since the socket is blocking the underlying accept
call blocks forever and since we have disabled the thread cancellation the
process deadlocks indefinitely. Note that before the previous commit the thread
could be cancelled even though the thread was blocked in the underlying accept()
call.

To avoid this we now switch to the much simpler (and more efficient) poll()
function and ensure we check the return value before calling accept.

fixes: commit 1b0ebd7 ("ENGBL-1307: Improve thread cancellation logic exposed in tests")

Change-Id: Ia923a593a59d267072248bffbf186068a49a110f
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Matt Bennett <matt.bennett@alliedtelesis.co.nz>