ENGBL-1307: Support TCP transport type with broadcast client
Currently the broadcast client code assumes the services are only ever
run via TIPC transports. Relax this assumption so that it can now be
used with TCP transports also. Note that the existing checks and event
generation is still only done for TIPC transports. This is okay for now,
but will be updated once TIPC unicast support is completely removed from
CMSG.

Note that we no longer connect the client before adding it to the
composite client. This is because for TCP servers we freebind the
socket hence the service listener learns of the server before the
route to the address is actually routable. This is really only a
problem for servers which start extremely early in AW+ (i.e. platform).
Therefore we simply delay the connect until the first API is called.
Since we listen for connections in a separate thread the need to
worry about when to connect is no longer required.

Change-Id: I582eefede80e6c453647c92fe65a06b2a2510592
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>