ENGBL-1307: Change composite client ipv4 lookup to take struct in_addr
This makes it obvious that the integer passed must be in network byte
order. Also taking struct in_addr will make the usages of this function
slightly cleaner as this is the type that the stack library returns.

Note that the struct only has a single uint32_t field hence there is
no loss in copying the struct versus taking a pointer.

Change-Id: Ie42c4772bb9c0a522bef6ac11febb666b40dd4dc
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>