ENGBL-921: Add forwarding client
Add a new client type that can be used instead of a user defined
transport when the user simply wants to forward the encoded messages
over some other medium. This avoids the need for the user to have
specific knowledge of how their transport must operate. Instead they
now simply get given the bytes to forward.

Change-Id: Ic606f7e89fef3bf82dfcb5a04f3e8b110b905111
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>