ENGBL-921: Delete user defined transport functionality
This code is no longer used and it is expected if any additional behaviour
is required of CMSG client/servers then that be added as first class
support to the CMSG library in the future. This avoids the need to
expose the internals of the library to applications and reduces coupling
to them.

Change-Id: I3c744c24ff99afd95716bad344e37c6f8975c9b9
Reviewed-by: John Thompson <john.thompson@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>