CR-71894: Add function to allow addition of unix client to broadcast client
Similar to adding a loopback client to the broadcast client allow the
user to add a unix client. This is currently a corner case usage of the
CMSG library, therefore this approach was chosen rather than making the
broadcast client also automatically connect to unix servers (rather than
currently just tcp).

Change-Id: I7a1ab4157480c0cb268a444e5634948b344c2ed4
Reviewed-by: Luuk Paulussen <luuk.paulussen@alliedtelesis.co.nz>
Reviewed-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>
Signed-off-by: Kyeong Yoo <kyeong.yoo@alliedtelesis.co.nz>